Exercises04/exercises-4-unit-tests/exercises-4-unit-tests/core/exercise1/FixedCapacityStackOfStrings.java
package core.exercise1;

public class FixedCapacityStackOfStrings {
    // holds the items
    private String[] a;
    // number of items in stack
    private int N;

    // create an empty stack with given capacity
    public FixedCapacityStackOfStrings(int capacity) {
        a = new String[capacity];
        N = 0;
    }

    public boolean isEmpty() {
        return N == 0;
    }

    public boolean isFull() {
        return N == a.length;
    }

    public int size() {
        return N;
    }

    public void push(String item) {
        a[N++] = item;
    }

    public String pop() {
        return a[--N];
    }

    public String peek() {
        return a[N - 1];
    }

} 


